Fix #202: Unify *Inks structs and extract ink builder
Replace four identical per-viz Inks structs with shared vizInks.
Extract buildMetricInk and helpers to ink_builder.go.
Remove redundant radialMetricValue wrapper.
