+++ cmd/codeviz/spiral_canvas.go
@@ -25,12 +25,6 @@ const (
 	spiralTrackMinSteps = 500
 )
 
-// spiralInks holds the Ink instances for a spiral render pass.
-type spiralInks struct {
-	fill   canvas.Ink
-	border canvas.Ink
-}
-
 // buildSpiralInks creates fill and border inks from aggregated time-bucket data.
 func buildSpiralInks(
 	buckets []spiral.TimeBucket,
@@ -38,8 +32,8 @@ func buildSpiralInks(
 	fillPaletteName palette.PaletteName,
 	borderMetric metric.Name,
 	borderPaletteName palette.PaletteName,
-) spiralInks {
-	inks := spiralInks{
+) vizInks {
+	inks := vizInks{
 		fill:   canvas.FixedInk(spiralDefaultFill),
 		border: canvas.FixedInk(spiralDefaultBorder),
 	}
@@ -112,7 +106,7 @@ func renderSpiralToCanvas(
 	layout spiral.SpiralLayout,
 	buckets []spiral.TimeBucket,
 	width, height int,
-	inks spiralInks,
+	inks vizInks,
 ) *canvas.Canvas {
 	cv := canvas.NewCanvas(width, height)
 
@@ -175,7 +169,7 @@ func addSpiralDiscs(
 	cv *canvas.Canvas,
 	nodes []spiral.SpiralNode,
 	buckets []spiral.TimeBucket,
-	inks spiralInks,
+	inks vizInks,
 ) {
 	for i, n := range nodes {
 		if n.DiscRadius <= 0 {
