+++ cmd/codeviz/viz_inks.go
@@ -1,0 +1,9 @@
+package main
+
+import "github.com/theunrepentantgeek/code-visualizer/internal/canvas"
+
+// vizInks holds the Ink instances for any visualization render pass.
+type vizInks struct {
+	fill   canvas.Ink
+	border canvas.Ink
+}
